requirements.txt
annotated-types==0.6.0
certifi==2023.7.22
charset-normalizer==3.3.2
click==8.1.7
dependency-injector==4.41.0
et-xmlfile==1.1.0
idna==3.4
mysql-connector-python==8.3.0
numpy==1.26.1
openpyxl==3.1.2
pandas==2.1.2
pydantic==2.4.2
pydantic_core==2.10.1
pymssql==2.2.10
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
requests==2.31.0
six==1.16.0
typer==0.9.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.7
